requirements.txt
asyncpg==0.29
fastapi==0.109.2
psycopg[pool,binary]==3.1.18
python-dotenv~=1.0
uvicorn[standard]~=0.27
